apps/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.14.0
# psycopg2-binary==2.9.10  # Not needed for SQLite
redis==5.2.0
pydantic==2.10.0
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
python-dotenv==1.0.1
celery==5.4.0
boto3==1.35.81
Pillow==11.0.0
openai==1.99.9
requests==2.32.3
pytest==8.3.4
pytest-asyncio==0.24.0
httpx>=0.25.0,<=0.27.0
weaviate-client>=4.9.0
numpy==1.26.4
tiktoken==0.8.0